Extract plate profile logic into separate module
